Fix Python client validation of map enum values
The generated Python client files for the cluster object had validation
of the featureAvailability map that checked if the map keys were among
the valid values for the enum used for map values. The validation check,
in all model classes and in the root template snippet causing the
ill-aimed validation, is updated to check the map values instead of the
map keys for valid values.

In OpenAPI, maps always have string keys. Values may be of just about
any type. Swagger code generation logic represents the OpenAPI
"additionalProperties" for a map, which talks about its value type, as
the map's "inner" type and therefore the values as its "items".
Furthermore, Swagger determines if a map is "enumerated" by looking at
its inner type / items. Therefore, it's the map values that are to be
checked as enumerations, not the keys.